refactor(eval): invoke CLI via `python -m strands_env.eval`, drop console script
Adopt the ML module-launcher convention (cf. `python -m sglang.launch_server`)
and make the eval subsystem own its CLI end-to-end:
- cli/eval.py -> eval/cli.py (eval_group: list/run); the cli/ package is removed.
- Add eval/__main__.py so `python -m strands_env.eval run ...` / `list` work.
- Drop the strands-env console script ([project.scripts]) — no PATH/console-script
  fragility in cluster/container runs; one consistent invocation.
